Move character events
