+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/PathTask.kt
@@ -3,7 +3,7 @@ package world.gregs.voidps.engine.client.update
 import world.gregs.voidps.engine.client.update.iterator.TaskIterator
 import world.gregs.voidps.engine.entity.character.Character
 import world.gregs.voidps.engine.entity.character.CharacterList
-import world.gregs.voidps.engine.entity.character.MoveStop
+import world.gregs.voidps.engine.entity.character.event.MoveStop
 import world.gregs.voidps.engine.entity.character.move.Path
 import world.gregs.voidps.engine.path.PathFinder
 import world.gregs.voidps.engine.path.PathResult
