+++ game/src/main/kotlin/world/gregs/voidps/bot/navigation/Navigation.kts
@@ -2,7 +2,7 @@ package world.gregs.voidps.bot.navigation
 
 import world.gregs.voidps.engine.action.ActionFinished
 import world.gregs.voidps.engine.action.ActionType
-import world.gregs.voidps.engine.entity.character.MoveStop
+import world.gregs.voidps.engine.entity.character.event.MoveStop
 import world.gregs.voidps.engine.entity.character.player.Bot
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.network.instruct.InteractObject
