+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/Death.kts
@@ -1,6 +1,6 @@
 import world.gregs.voidps.engine.action.ActionType
 import world.gregs.voidps.engine.entity.character.Character
-import world.gregs.voidps.engine.entity.character.Death
+import world.gregs.voidps.engine.entity.character.event.Death
 import world.gregs.voidps.engine.entity.character.player.skill.CurrentLevelChanged
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.event.on
