+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/DeathPlayer.kts
@@ -3,14 +3,17 @@ import kotlinx.coroutines.withContext
 import world.gregs.voidps.engine.action.ActionType
 import world.gregs.voidps.engine.action.action
 import world.gregs.voidps.engine.entity.*
-import world.gregs.voidps.engine.entity.character.*
-import world.gregs.voidps.engine.entity.character.Death
+import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.clearAnimation
 import world.gregs.voidps.engine.entity.character.contain.Container
 import world.gregs.voidps.engine.entity.character.contain.equipment
 import world.gregs.voidps.engine.entity.character.contain.inventory
+import world.gregs.voidps.engine.entity.character.event.Death
+import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.move.move
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.isAdmin
+import world.gregs.voidps.engine.entity.character.setAnimation
 import world.gregs.voidps.engine.entity.item.FloorItems
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.engine.map.Tile
