+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/npc/move/Walk.kts
@@ -1,10 +1,14 @@
 import world.gregs.voidps.engine.action.ActionFinished
 import world.gregs.voidps.engine.entity.*
-import world.gregs.voidps.engine.entity.character.*
-import world.gregs.voidps.engine.entity.character.CantReach
+import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.event.CantReach
+import world.gregs.voidps.engine.entity.character.event.MoveStop
+import world.gregs.voidps.engine.entity.character.event.Moving
+import world.gregs.voidps.engine.entity.character.face
 import world.gregs.voidps.engine.entity.character.move.Path
 import world.gregs.voidps.engine.entity.character.move.cantReach
 import world.gregs.voidps.engine.entity.character.move.withinDistance
+import world.gregs.voidps.engine.entity.character.watch
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.engine.path.strat.TileTargetStrategy
