+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/ForceMovement.kts
@@ -1,5 +1,5 @@
 import world.gregs.voidps.engine.entity.character.Character
-import world.gregs.voidps.engine.entity.character.Moved
+import world.gregs.voidps.engine.entity.character.event.Moved
 import world.gregs.voidps.engine.entity.contains
 import world.gregs.voidps.engine.entity.remove
 import world.gregs.voidps.engine.event.on
