+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/music/Music.kts
@@ -9,7 +9,7 @@ import world.gregs.voidps.engine.client.variable.addVar
 import world.gregs.voidps.engine.client.variable.hasVar
 import world.gregs.voidps.engine.client.variable.sendVar
 import world.gregs.voidps.engine.entity.Registered
-import world.gregs.voidps.engine.entity.character.Moved
+import world.gregs.voidps.engine.entity.character.event.Moved
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.definition.EnumDefinitions
 import world.gregs.voidps.engine.entity.get
