+++ game/src/main/kotlin/world/gregs/voidps/world/map/CantReach.kts
@@ -1,5 +1,5 @@
 import world.gregs.voidps.engine.action.ActionType
-import world.gregs.voidps.engine.entity.character.CantReach
+import world.gregs.voidps.engine.entity.character.event.CantReach
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.cantReach
 import world.gregs.voidps.engine.event.on
